Changed total price to now be a dynamically calculated value following QTY * UP

+++ config/receipt_parser/models.py
@@ -39,7 +39,7 @@ class Items(models.Model):
         related_name="rel_items_category_id_fk",
     )
     item_name = models.CharField(max_length=100, unique=False, null=False)
-    price = models.IntegerField(null=False)
+    price = models.FloatField(null=False)
     item_description = models.CharField(max_length=255, unique=False, null=True)
     item_insert_datetime = models.DateTimeField(auto_now_add=True)
